Perl_valid_utf8_to_uvchr() unroll loop for speed
This unrolls a loop to speed things up.  I had previously found no
difference in speed, but others have found a significant change, and
it's easy enough to do.

I used macro definitions to make the inline code look cleaner, and show
the patterns in the macros better, but more importantly most of them
will be used in a future commit elsewhere in this file.