utf8.c: Reorder uvoffuni_to_utf8() to avoid branches
Prior to this commit, this function had a branch at the beginning to
look for problematic code points that may have to be warned about or
disallowed.  This moves the code out of there to later in the routine to
where we know we have a range of code points that could contain the
problematic code points, but not in ranges that don't contain them.
This should save execution branching.