Skip casing for some non-cased scripts
Characters whose upper, lower, title, or fold case differ from the
character itself amount to just 1.5% of the assigned Unicode characters,
and this percentage falls with each new Unicode release, as almost all
cased scripts have already been encoded.  But a lot of code is written
assuming a cased language, such as calling uc() or lcfirst(), or doing
qr//i.  When such code is run on a non-cased language, the work expended
in doing the casing is wasted.  And casing is expensive.  But finding
out if a character is cased or not is nearly as expensive, so one might
as well just do the casing.  However, the Unicode code space is
organized so that there are some long stretches of contiguous code
points that aren't cased.  By adding tests to see if the input code
point is in just a few of these ranges, we can quickly rule casing out
for most of the non-cased scripts that are of commercial use today, at
essentially no expense to handling the more common cased scripts.
Testing for just 3 ranges in Plane 0 of Unicode (where most of the code
points in common use today reside) allows us to skip doing casing for
more than 82% of code points in the plane, including the following
languages: Arabic, Chinese, Hebrew, Japanese, Korean, Thai, and the
major scripts of India.  No longer is a swash generated when trying to
case one of these, so runtime memory usage is decreased.

(It should be noted that some of these languages have characters
scattered in other areas, because the original allocation for them
turned out to be not large enough.  When changing the case of these
other characters, the lookups won't be skippped.  But that original
allocation included all the characters in current common use, so these
other characters are comparatively rare.)

When this command is run

    blead Porting/bench.pl --raw --perlargs="-Ilib -X" --benchfile=/home/khw/perl/perf_for_casing ./perl=before_this_commit ../ebcdic/perl=after

and perf_for_casing contains:
    [
        'string::casing::greek' => {
            desc    => 'should be no change',
            setup   => 'my $a = "\x{3B1}"',  # GREEK SMALL LETTER ALPHA
            code    => 'uc($a)'
        },

        'string::casing::hebrew' => {
            desc    => 'yes swash vs no swash',
            setup   => 'my $a = "\x{5D0}"',  # HEBREW LETTER ALEF
            code    => 'uc($a)'
        },
    ];

These are the results:

    string::casing::greek
    should be no change

           before_this_commit    after
           ------------------ --------
        Ir             1638.0   1638.0
        Dr              707.0    707.0
        Dw              200.0    200.0
      COND              185.0    185.0
       IND                8.0      8.0

    COND_m                3.0      5.0
     IND_m                2.0      2.0

     Ir_m1                0.0      0.0
     Dr_m1                0.0      0.0
     Dw_m1                0.0      0.0

     Ir_mm                0.0      0.0
     Dr_mm                0.0      0.0
     Dw_mm                0.0      0.0

    string::casing::hebrew
    yes swash vs no swash

           before_this_commit    after
           ------------------ --------
        Ir             1616.0    678.0
        Dr              701.0    264.0
        Dw              190.0     81.0
      COND              192.0     91.0
       IND                8.0      5.0

    COND_m                3.0      1.0
     IND_m                2.0      2.0

     Ir_m1                0.0      0.0
     Dr_m1                0.0      0.0
     Dw_m1                0.0      0.0

     Ir_mm                0.0      0.0
     Dr_mm                0.0      0.0
     Dw_mm                0.0      0.0