Skip casing for high code points
As discussed in the previous two commits, most code points in Unicode
don't change if upper-, or lower-cased, etc.  In fact as of Unicode
v8.0, 93% of the available code points are above the highest one that
does change.

This commit skips trying to case these 93% of Unicode code points.  A
regen/ script keeps track of the max changing one in the current Unicode
release, and skips casing for the higher ones.  Thus currently, casing
emoji will be skipped.

If the command

    blead Porting/bench.pl --raw --perlargs="-Ilib -X" --benchfile=/home/khw/perl/perf_for_casing ./perl=before_this_commit ../ebcdic/perl=after

is run and the file 'perf_for_casing' contains

    [
        'string::casing::emoji' => {
            desc    => 'yes swash vs no swash',
            setup   => 'my $a = "\x{1F570}"',  # MANTELPIECE CLOCK
            code    => 'uc($a)'
        },
    ];

the following results are obtained:

    string::casing::emoji
    yes swash vs no swash

           before_this_commit    after
           ------------------ --------
        Ir             1651.0    696.0
        Dr              718.0    270.0
        Dw              191.0     81.0
      COND              193.0     95.0
       IND                9.0      5.0

    COND_m                5.0      1.0
     IND_m                4.0      2.0

     Ir_m1                0.0      0.0
     Dr_m1                0.0      0.0
     Dw_m1                0.0      0.0

     Ir_mm                0.0      0.0
     Dr_mm                0.0      0.0
     Dw_mm                0.0      0.0