Don't try to case change surrogates, above-Unicodes
Changing the case (upper, lower, title, fold) of surrogate code points
and non-Unicode code points always yields the original, so there is no
need to actually try it.  And trying it is slow and creates swashes,
which uses up runtime memory.  We test for these anyway, so at the cost
of just two gotos and a label, we can skip all that work and potential
memory use.  This is worth doing even though such usage will be rare in
practice.

Running the following command

    blead Porting/bench.pl --perlargs="-Ilib -X" --benchfile=/home/khw/perl/perf_for_casing ./perl=before_this_commit ../ebcdic/perl=after

where perf_for_casing has these contents:

    [
        'string::casing' => {
            desc    => 'yes swash vs no swash',
            setup   => 'my $a = "\x{110000}"',
            code    => 'uc($a)'
        },
    ];

yields this output (higher numbers are better)

       before_this_commit  after
        ------------------ ------
     Ir             100.00 232.92
     Dr             100.00 261.57
     Dw             100.00 223.60
   COND             100.00 210.53
    IND             100.00 180.00

 COND_m             100.00 500.00
  IND_m             100.00 200.00

  Ir_m1             100.00 100.00
  Dr_m1             100.00 100.00
  Dw_m1             100.00 100.00

  Ir_mm             100.00 100.00
  Dr_mm             100.00 100.00
  Dw_mm             100.00 100.00